--- nextjs_frontend/src/app/importar/page.tsx
@@ -1,142 +1,0 @@
-"use client";
-
-import { useState } from "react";
-
-import Button from "@/components/Button/button";
-import { useDropzone } from "react-dropzone";
-import Papa from "papaparse";
-import { serverURL } from "@/config";
-import { useRouter } from "next/navigation";
-
-export default function ImportPage() {
-  const router = useRouter();
-  const [droppedFiles, setDroppedFiles] = useState<any[]>([]);
-
-  const validateCSVHeaders = (file: any) => {
-    return new Promise((resolve, reject) => {
-      Papa.parse(file, {
-        header: true,
-        skipEmptyLines: true,
-        complete: (result: any) => {
-          const headers = result.meta.fields;
-          const requiredHeaders = [
-            "cpf",
-            "nome",
-            "valorDivida",
-            "nomeCondominio",
-          ];
-
-          const isValid = requiredHeaders.every((header) =>
-            headers.includes(header)
-          );
-
-          if (isValid) {
-            resolve(true);
-          } else {
-            reject(
-              "Invalid headers in the CSV file. Please make sure the headers are 'cpf', 'nome', 'valorDivida', and 'nomeCondominio'."
-            );
-          }
-        },
-        error: (error) => {
-          reject("Error parsing CSV file.");
-        },
-      });
-    });
-  };
-
-  const { getRootProps, getInputProps } = useDropzone({
-    onDrop: async (acceptedFiles) => {
-      const csvRegex = new RegExp(".csv$");
-      const isCSV = acceptedFiles.every((file) => csvRegex.test(file.name));
-
-      if (!isCSV) {
-        alert("Arquivo inválido. Por favor, selecione apenas arquivos .csv");
-        return;
-      }
-
-      const hasValidHeaders = await Promise.all(
-        acceptedFiles.map(async (file) => {
-          try {
-            await validateCSVHeaders(file);
-            return true;
-          } catch (error) {
-            console.error(`Erro processando o arquivo ${file.name}: ${error}`);
-            return false;
-          }
-        })
-      );
-
-      if (hasValidHeaders.every((isValid) => isValid)) {
-        setDroppedFiles(acceptedFiles);
-      } else {
-        alert(
-          "Arquivo inválido. Por favor, verifique se todos os arquivos possuem as colunas 'cpf', 'nome', 'valorDivida' e 'nomeCondominio'."
-        );
-      }
-    },
-  });
-
-  const handleSubmit = async () => {
-    const formData = new FormData();
-    droppedFiles.forEach((file) => formData.append("files", file));
-
-    const response = await fetch(`${serverURL}/api/tenants/import`, {
-      method: "POST",
-      body: formData,
-    });
-
-    if (response.ok) {
-      alert("Devedores adicionados com sucesso!");
-      setDroppedFiles([]);
-      router.push("/tenants");
-    } else {
-      alert("Erro ao enviar os arquivos.");
-    }
-  };
-
-  return (
-    <div className="containerLayout">
-      <div className="my-10">
-        <h1 className="text-4xl font-bold leading-10">
-          Adicione devedores por meio de arquivos .csv:
-        </h1>
-      </div>
-
-      <div
-        {...getRootProps()}
-        className="mt-4 p-4 border border-dashed cursor-pointer"
-        style={{
-          display: "flex",
-          flexDirection: "column",
-          alignItems: "center",
-          justifyContent: "center",
-          textAlign: "center",
-          border: "2px dashed #000",
-          height: "200px",
-        }}
-      >
-        <input {...getInputProps()} />
-        <p className="text-gray-600">
-          Solte seus arquivos aqui ou clique para selecionar.
-        </p>
-      </div>
-
-      {droppedFiles.length > 0 && (
-        <div className="flex items-center justify-between">
-          <div className="mt-4">
-            <h3 className="text-lg font-semibold">Arquivos:</h3>
-            <ul>
-              {droppedFiles.map((file) => (
-                <li key={file.name}>{file.name}</li>
-              ))}
-            </ul>
-          </div>
-          <div className="mt-4 self-start">
-            <Button onClick={handleSubmit}>Enviar</Button>
-          </div>
-        </div>
-      )}
-    </div>
-  );
-}
